+++ WhatsNew.md
@@ -1,5 +1,6 @@
 # Changes
 
-- Fix reputation display when friend has paragon mode (TWW: The Severed Threads' members)
+- Fix reputation icon positioning for renown with paragon (DF and TWW major factions, TWW Threads friends)
+- Add sorting to favorite faction popup (by faction ID descending)
 
 #### Complete Changelog